[architect] refactor: extract governor endpoints from api.go
Refs #6570 slice 2/5.\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
